Adição do Arma

+++ src/br/poli/sots/utils/serie/SerieYear.java
@@ -4,17 +4,17 @@ import java.util.LinkedList;
 import java.util.List;
 
 public class SerieYear {
-	public List<Integer> riverFlow;
+	public List<Double> riverFlow;
 	
 	public int year;
 	public int average;
 	
 	public SerieYear()
 	{
-		riverFlow = new LinkedList<Integer>();
+		riverFlow = new LinkedList<Double>();
 	}
 	
-	public SerieYear(int year, List<Integer> riverFlow){
+	public SerieYear(int year, List<Double> riverFlow){
 		this.year = year;
 		this.riverFlow = riverFlow;
 	}
